vs_platform/pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name        = "vs-agent-platform"
version     = "0.1.0"
description = "Vidya Sankalp Agentic Platform — FastAPI gateway, auth, rate limiting, prompt versioning"
readme      = "README.md"
requires-python = ">=3.11"

dependencies = [
    "vs-agent-core>=0.1.0",
    "langsmith==0.6.9",
    "langchain-core==1.2.22",
    "fastapi",
    "uvicorn[standard]",
    "starlette",
    "python-multipart",
    "PyJWT",
    "cryptography",
    "httpx",
    "boto3",
    "pydantic",
    "pydantic-settings",
    "tenacity",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-html",
    "httpx",
]

[tool.setuptools.packages.find]
where   = ["."]
include = ["vs_platform*"]

[project.urls]
Homepage   = "https://vidyasankalp.com"
Repository = "https://github.com/vidyasankalp/vs-agentic-platform"
